requirements.txt
python-telegram-bot==20.7
python-dotenv==1.0.0
aiohttp==3.9.1
flask==2.3.3
flask-cors==4.0.0
gunicorn==21.2.0